basic module refactor and structure
