+++ clabe/types/__init__.py
@@ -1,1 +1,6 @@
-from .clabes_pydantic_legacy import Clabe, validate_digits
+from ..helpers import is_pydantic_v1_installed
+
+if is_pydantic_v1_installed():
+    from .clabes_pydantic_v1 import Clabe, validate_digits
+else:
+    from .clabes import Clabe, validate_digits
