+++ clabe/types/clabes.py
@@ -1,0 +1,4 @@
+def validate_digits(v: str) -> str: ...
+
+
+class Clabe(str): ...
