+++ requirements.txt
@@ -1,1 +1,1 @@
-pydantic==1.9.0
+pydantic==1.10.0
