+++ setup.py
@@ -21,7 +21,7 @@ setuptools.setup(
     packages=setuptools.find_packages(),
     include_package_data=True,
     package_data=dict(clabe=['py.typed']),
-    install_requires=['pydantic>=1.4,<2.0'],
+    install_requires=['pydantic>=1.10,<3.0'],
     classifiers=[
         'Programming Language :: Python :: 3',
         'License :: OSI Approved :: MIT License',
